adding class currency calculator fun, fix bug in clsString

+++ level_9/currancy_opp/clsInputValidate.h
@@ -36,7 +36,7 @@ class clsInputValidate{
         }
         return Number;
     }
-    static double ReadFloatNumber(string Error = "Invalid number") {
+    static float ReadFloatNumber(string Error = "Invalid number") {
         float Number = 0;
         while (!(cin>> Number)) {
             cin.clear();
